fixed null image kitOrder

+++ app/src/main/java/com/example/rfidtab/service/response/user/UserInfoResponse.kt
@@ -3,8 +3,8 @@ package com.example.rfidtab.service.response.user
 import com.google.gson.annotations.SerializedName
 
 class UserInfoResponse(
-    @SerializedName("id") val id: Int,
-    @SerializedName("phone") val phone: Int,
+    @SerializedName("id") val id: Long,
+    @SerializedName("phone") val phone: Long,
     @SerializedName("email") val email: String,
     @SerializedName("fio") val fio: String,
     @SerializedName("roleName") val roleName: String,
